requirements.txt
accelerate
bitsandbytes
transformers
datasets
peft
trl
tiktoken
protobuf
sentencepiece
pillow
torchvision
openpyxl

